+++ src/schematics/SchematicDrivenLayout.jl
@@ -18,6 +18,7 @@ import Logging:
     shouldlog,
     with_logger
 import LoggingExtras: FormatLogger, MinLevelLogger, TeeLogger
+import StaticArrays
 
 import DeviceLayout:
     AbstractComponent,
@@ -83,6 +84,7 @@ export @component,
     components,
     create_component,
     default_parameters,
+    extract_parameter_set,
     facing,
     filter_parameters,
     find_components,
@@ -141,6 +143,7 @@ include("components/components.jl")
 include("components/compdef.jl")
 include("components/composite_components.jl")
 include("components/builtin_components.jl")
+include("parameter_set_extraction.jl")
 include("routes.jl")
 include("components/variants.jl")
 include("solidmodels.jl")